third_party/WebKit/LayoutTests/fast/dom/zoom-scroll-page-test-expected.txt

This tests that window.scrollBy respects the page's zoom and scale factor when scrolling.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

Testing at zoom -1 and scale 1.2
PASS window.pageYOffset is within 4 of 231
Testing at zoom -3 and scale 0.75
PASS window.pageYOffset is within 4 of 231
Testing at zoom -3 and scale 1
PASS window.pageYOffset is within 4 of 231
Testing at zoom -3 and scale 1.2
PASS window.pageYOffset is within 4 of 231
Testing at zoom -3 and scale 1.6
PASS window.pageYOffset is within 4 of 231
Testing at zoom -2 and scale 0.75
PASS window.pageYOffset is within 4 of 231
Testing at zoom -2 and scale 1
PASS window.pageYOffset is within 4 of 231
Testing at zoom -2 and scale 1.2
PASS window.pageYOffset is within 4 of 231
Testing at zoom -2 and scale 1.6
PASS window.pageYOffset is within 4 of 231
Testing at zoom -1 and scale 0.75
PASS window.pageYOffset is within 4 of 231
Testing at zoom -1 and scale 1
PASS window.pageYOffset is within 4 of 231
Testing at zoom -1 and scale 1.2
PASS window.pageYOffset is within 4 of 231
Testing at zoom -1 and scale 1.6
PASS window.pageYOffset is within 4 of 231
Testing at zoom 0 and scale 0.75
PASS window.pageYOffset is within 4 of 231
Testing at zoom 0 and scale 1
PASS window.pageYOffset is within 4 of 231
Testing at zoom 0 and scale 1.2
PASS window.pageYOffset is within 4 of 231
Testing at zoom 0 and scale 1.6
PASS window.pageYOffset is within 4 of 231
Testing at zoom 1 and scale 0.75
PASS window.pageYOffset is within 4 of 231
Testing at zoom 1 and scale 1
PASS window.pageYOffset is within 4 of 231
Testing at zoom 1 and scale 1.2
PASS window.pageYOffset is within 4 of 231
Testing at zoom 1 and scale 1.6
PASS window.pageYOffset is within 4 of 231
Testing at zoom 3 and scale 0.75
PASS window.pageYOffset is within 4 of 231
Testing at zoom 3 and scale 1
PASS window.pageYOffset is within 4 of 231
Testing at zoom 3 and scale 1.2
PASS window.pageYOffset is within 4 of 231
Testing at zoom 3 and scale 1.6
PASS window.pageYOffset is within 4 of 231
Testing at zoom 5 and scale 0.75
PASS window.pageYOffset is within 4 of 231
Testing at zoom 5 and scale 1
PASS window.pageYOffset is within 4 of 231
Testing at zoom 5 and scale 1.2
PASS window.pageYOffset is within 4 of 231
Testing at zoom 5 and scale 1.6
PASS window.pageYOffset is within 4 of 231
PASS successfullyParsed is true

TEST COMPLETE
